Move focusing logic to the Screen level
